# Subtile decoding: memory use reduction and perf improvements